Copy constants.py file from streamlit app
