+++ flask_app/legacy/populate_db.py
@@ -7,31 +7,8 @@ import itertools
 import flask_app.legacy.db_functions as db
 from flask_app.legacy.yml_functions import read_yml
 from flask_app.legacy.parse_raw_data import get_techniques_metabolites, get_measures_growth, get_measures_reads, get_measures_counts, get_replicate_metadata
+from flask_app.legacy.constants import GrowthTechniques, Vessels, LOCAL_DIRECTORY_YAML
 
-# TODO Temporary:
-root_folder = os.path.abspath(os.path.dirname(__file__) + '/../../')
-
-LOCAL_DIRECTORY           = os.path.join(root_folder, "")
-LOCAL_DIRECTORY_APP       = os.path.join(LOCAL_DIRECTORY, "app")
-LOCAL_DIRECTORY_TEMPLATES = os.path.join(LOCAL_DIRECTORY_APP, "templates")
-LOCAL_DIRECTORY_YAML      = os.path.join(LOCAL_DIRECTORY_TEMPLATES, "yaml_templates")
-
-class GrowthTechniques:
-    def __init__(self):
-        self.od = "Optical Density"
-        self.plates = "Plate Counts"
-        self.plates_ps = "Plate Counts (per species)"
-        self.fc = "Flow Cytometry"
-        self.fc_ps = "Flow Cytometry (per species)"
-        self.rna = "16S rRNA-seq"
-
-class Vessels:
-    def __init__(self):
-        # 'Bottles', 'Agar-plates', 'Well-plates', 'mini-bioreactors'
-        self.bottles = "Bottles"
-        self.agar_plates = "Agar plates"
-        self.well_plates = "Well plates"
-        self.mini_react = "mini bioreactors"
 
 def save_submission_to_database(conn, submission, data_template):
     """
